+++ src/main/kotlin/io/unthrottled/amii/onboarding/UpdateNotification.kt
@@ -24,6 +24,7 @@ private fun buildUpdateMessage(updateAsset: String): String =
       <ul>
         <li>Added Integrated Discreet Mode</li>
         <li>2021.3 Build Support</li>
+        <li>Handling vertical overflow better in settings UI</li>
       </ul>
       <br>See the <a href="https://github.com/ani-memes/AMII#documentation">documentation</a> for features, usages, and configurations.
       <br>The <a href="https://github.com/ani-memes/AMII/blob/master/CHANGELOG.md">changelog</a> is available for more details.