+++ src/main/kotlin/io/unthrottled/amii/memes/MemePanel.kt
@@ -61,6 +61,10 @@ import javax.swing.MenuElement
 import javax.swing.SwingUtilities
 import kotlin.math.max
 
+enum class ClickEvent {
+  LEFT, RIGHT, UNKNOWN
+}
+
 enum class PanelDismissalOptions {
   FOCUS_LOSS, TIMED;
 
@@ -159,6 +163,7 @@ class MemePanel(
     )
   }
 
+  @Suppress("ComplexMethod") // cuz I said so.
   private fun createMouseListener(): AWTEventListener {
     var clickedInside = false
     return AWTEventListener { event ->
@@ -179,7 +184,13 @@ class MemePanel(
           } else if (wasInside) {
             fadeoutAlarm.cancelAllRequests()
             clickedInside = true
-            this.lifecycleListener.onClick()
+            this.lifecycleListener.onClick(
+              when {
+                SwingUtilities.isLeftMouseButton(event) -> ClickEvent.LEFT
+                SwingUtilities.isRightMouseButton(event) -> ClickEvent.RIGHT
+                else -> ClickEvent.UNKNOWN
+              }
+            )
           }
         }
       } else if (