+++ docs/RELEASE-NOTES.md
@@ -8,3 +8,8 @@
 - A `Clear Memes` action just in case I didn't actually fix invulnerable memes.
   - Available via `Tools -> AMII Options -> Clear Memes`
 - 2021.2 build support!
+
+### Changed
+
+- Assets that get downloaded in the background. Peferred assets will only be downloaded, that way I'm not paying for
+  users to download assets that they won't see.