+++ docs/RELEASE-NOTES.md
@@ -2,3 +2,4 @@
 
 - `Discreet Mode`, when enabled MIKU will clear and not show _any_ anime in the IDE. Also, the mood in the status bar
   will temporarily hide as well.
+- Added 2021.3 Build support. Plugin only supports 2020.3+ builds now.