+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,7 @@
 ### Added
 
+- The ability to view information about the source of a meme on click.
+- `Show Previous Meme`: an action that shows the previously shown meme.
 - `Discreet Mode`, when enabled MIKU will clear and not show _any_ anime in the IDE. Also, the mood in the status bar
   will temporarily hide as well.
 - Added 2021.3 Build support. Plugin only supports 2020.3+ builds now.