+++ docs/RELEASE-NOTES.md
@@ -1,3 +1,8 @@
+### Added
+
+- 2022.1 Build Support
+- The ability to right-click a meme view to the meme's asset source page on https://amii-assets.unthrottled.io/
+
 ### Fixed
 
 - Images not showing up for users with an `'` in their file path.